refactor: renamed useInfiniteDepartmentsList to useDepartmentsList
